DiarizationLM/requirements.txt
numpy
scipy
tensorflow-text
absl-py
openai
datasets
tqdm
